Move ManifestCore into manifest module
