+++ slatedb/src/db_state.rs
@@ -1,14 +1,11 @@
 use crate::bytes_range::BytesRange;
-use crate::checkpoint::Checkpoint;
 use crate::config::CompressionCodec;
 use crate::error::SlateDBError;
 use crate::manifest::Manifest;
 use crate::mem_table::{ImmutableMemtable, KVTable, WritableKVTable};
 use crate::reader::DbStateReader;
-use crate::seq_tracker::SequenceTracker;
 use crate::wal_id::WalIdStore;
 use bytes::Bytes;
-use log::debug;
 use serde::Serialize;
 use slatedb_txn_obj::DirtyObject;
 use std::collections::VecDeque;
@@ -17,8 +14,9 @@ use std::ops::Bound::{Excluded, Included, Unbounded};
 use std::ops::{Bound, Range, RangeBounds};
 use std::sync::Arc;
 use ulid::Ulid;
-use uuid::Uuid;
 use SsTableId::{Compacted, Wal};
+
+pub(crate) use crate::manifest::ManifestCore;
 
 /// A handle to an SSTable — the physical SST on storage.
 #[derive(Clone, PartialEq, Serialize)]
@@ -469,117 +467,6 @@ pub(crate) struct COWDbState {
 impl COWDbState {
     pub(crate) fn core(&self) -> &ManifestCore {
         &self.manifest.value.core
-    }
-}
-
-/// Represents an immutable in-memory view of .manifest file that is suitable
-/// to expose to end-users.
-#[derive(Clone, PartialEq, Serialize, Debug)]
-pub struct ManifestCore {
-    /// Flag to indicate whether initialization has finished. When creating the initial manifest for
-    /// a root db (one that is not a clone), this flag will be set to true. When creating the initial
-    /// manifest for a clone db, this flag will be set to false and then updated to true once clone
-    /// initialization has completed.
-    pub initialized: bool,
-
-    /// The last compacted l0 SstView ID.
-    pub last_compacted_l0_sst_view_id: Option<Ulid>,
-
-    /// The SST ID of the last compacted L0. In V2, view IDs differ from SST IDs,
-    /// but V1 only stores SST IDs. This field preserves the SST ID so that a
-    /// V1-encoded manifest can correctly reference the compacted L0.
-    pub last_compacted_l0_sst_id: Option<Ulid>,
-
-    /// A list of the L0 SST views that are valid to read in the `compacted` folder.
-    pub l0: VecDeque<SsTableView>,
-
-    /// A list of the sorted runs that are valid to read in the `compacted` folder.
-    pub compacted: Vec<SortedRun>,
-
-    /// The next WAL SST ID to be assigned when creating a new WAL SST. The manifest FlatBuffer
-    /// contains `wal_id_last_seen`, which is always one less than this value.
-    pub next_wal_sst_id: u64,
-
-    /// the WAL ID after which the WAL replay should start. Default to 0,
-    /// which means all the WAL IDs should be greater than or equal to 1.
-    /// When a new L0 is flushed, we update this field to the recent
-    /// flushed WAL ID.
-    pub replay_after_wal_id: u64,
-
-    /// the `last_l0_clock_tick` includes all data in L0 and below --
-    /// WAL entries will have their latest ticks recovered on replay
-    /// into the in-memory state.
-    pub last_l0_clock_tick: i64,
-
-    /// it's persisted in the manifest, and only updated when a new L0
-    /// SST is created in the manifest.
-    pub last_l0_seq: u64,
-
-    /// Minimum sequence number across all recent in-memory snapshots. The compactor
-    /// needs this to determine whether it's safe to drop duplicate key writes. If a
-    /// recent snapshot still references an older version of a key, it should not be
-    /// recycled. This field is updated when a new L0 is flushed.
-    pub recent_snapshot_min_seq: u64,
-
-    /// A sequence tracker that maps sequence numbers to timestamps as defined in
-    /// RFC-0012.
-    pub sequence_tracker: SequenceTracker,
-
-    /// A list of checkpoints that are currently open.
-    pub checkpoints: Vec<Checkpoint>,
-
-    /// The URI of the object store dedicated specifically for WAL, if any.
-    pub wal_object_store_uri: Option<String>,
-}
-
-impl ManifestCore {
-    pub(crate) fn new() -> Self {
-        Self {
-            initialized: true,
-            last_compacted_l0_sst_view_id: None,
-            last_compacted_l0_sst_id: None,
-            l0: VecDeque::new(),
-            compacted: vec![],
-            next_wal_sst_id: 1,
-            replay_after_wal_id: 0,
-            last_l0_clock_tick: i64::MIN,
-            last_l0_seq: 0,
-            checkpoints: vec![],
-            wal_object_store_uri: None,
-            recent_snapshot_min_seq: 0,
-            sequence_tracker: SequenceTracker::new(),
-        }
-    }
-
-    pub(crate) fn new_with_wal_object_store(wal_object_store_uri: Option<String>) -> Self {
-        let mut this = Self::new();
-        this.wal_object_store_uri = wal_object_store_uri;
-        this
-    }
-
-    pub(crate) fn init_clone_db(&self) -> ManifestCore {
-        let mut clone = self.clone();
-        clone.initialized = false;
-        clone.checkpoints.clear();
-        clone
-    }
-
-    pub(crate) fn log_db_runs(&self) {
-        let l0s: Vec<_> = self.l0.iter().map(|l0| l0.estimate_size()).collect();
-        let compacted: Vec<_> = self
-            .compacted
-            .iter()
-            .map(|sr| (sr.id, sr.estimate_size()))
-            .collect();
-        debug!("DB Levels:");
-        debug!("-----------------");
-        debug!("{:?}", l0s);
-        debug!("{:?}", compacted);
-        debug!("-----------------");
-    }
-
-    pub(crate) fn find_checkpoint(&self, checkpoint_id: Uuid) -> Option<&Checkpoint> {
-        self.checkpoints.iter().find(|c| c.id == checkpoint_id)
     }
 }
